% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_bb}
\alias{get_bb}
\title{Retrieve a bounding box}
\usage{
get_bb(reg = "gom")
}
\arguments{
\item{reg}{chr, the name of the region}
}
\value{
a bounding box ("bbox" class)
}
\description{
Retrieve a bounding box
}
