% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{list_images}
\alias{list_images}
\title{List images}
\usage{
list_images(
  what = c("wrapped", "daily")[1],
  path = system.file("extdata", package = "calfinforecast")
)
}
\arguments{
\item{what}{chr either "wrapped" (default) or "daily"}

\item{path}{chr the path to search}

\item{chr}{vector of files (possibly named)}
}
\description{
List images
}
