% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_coastline}
\alias{read_coastline}
\title{Read a coastline}
\usage{
read_coastline(
  filename = system.file("extdata/coastline.Rds", package = "calfinforecast")
)
}
\arguments{
\item{filename}{str, the filename}
}
\value{
sfc (geometry) object
}
\description{
Made with Natural Earth. Free vector and raster map data at naturalearthdata.com.
}
