% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_dmr_gzmp}
\alias{read_dmr_gzmp}
\title{Read a DMR GZMP site metadata}
\usage{
read_dmr_gzmp(
  filename = system.file("extdata/dmr_gzmp_sites.Rds", package = "calfinforecast"),
  form = c("data.frame", "sf")[2]
)
}
\arguments{
\item{filename}{str, the filename}

\item{form}{chr, one of "data.frame" or "sf" (default)}
}
\value{
sf object
}
\description{
Read a DMR GZMP site metadata
}
