% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_dmr_zones}
\alias{read_dmr_zones}
\title{Read a DMR lobster zones}
\usage{
read_dmr_zones(
  filename = system.file("extdata/dmr_lobster_zones.Rds", package = "calfinforecast")
)
}
\arguments{
\item{filename}{str, the filename}
}
\value{
sf object
}
\description{
Read a DMR lobster zones
}
