% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{save_graphics}
\alias{save_graphics}
\title{Save graphics as PNGs}
\usage{
save_graphics(x = plot_forecast(), path = data_path())
}
\arguments{
\item{x}{either a ggplot object (facte wrapped by time) or a list
of daily graphics}

\item{path}{the path to write to}
}
\description{
Save graphics as PNGs
}
