% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_config}
\alias{write_config}
\alias{read_config}
\title{Read and write the config list}
\usage{
write_config(x, filename = data_path("config.yaml"))

read_config(
  filename = system.file("extdata/config.yaml", package = "calfinforecast")
)
}
\arguments{
\item{x}{configuration list}

\item{filename}{chr then name of the file}
}
\value{
for \code{write_config} and \code{read_config} a configuration list
}
\description{
Read and write the config list
}
