% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_raster}
\alias{write_raster}
\alias{read_raster}
\title{Read and write raster data}
\usage{
write_raster(x, filename = data_path("data.Rds"))

read_raster(
  filename = system.file("extdata/data.Rds", package = "calfinforecast")
)
}
\arguments{
\item{x}{stars object}

\item{filename}{chr then name of the file}
}
\value{
for \code{write_raster} and \code{read_raster} a stars object
}
\description{
Read and write raster data
}
