% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calfinforecast-package.R
\docType{package}
\name{calfinforecast}
\alias{calfinforecast-package}
\alias{calfinforecast}
\title{calfinforecast}
\description{
Serves a subset of ecopmo calfin predictive maps
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BigelowLab/calfinforecast}
}

}
\keyword{internal}
