% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{data_path}
\alias{data_path}
\title{Retrieve the package data path}
\usage{
data_path(
  ...,
  root =
    "/mnt/s1/projects/ecocast/corecode/R/ecopmo_forecast/calfin_forecast/inst/extdata"
)
}
\arguments{
\item{...}{file path segments to append to the root path}

\item{root}{the root data path}
}
\value{
package data path
}
\description{
Retrieve the package data path
}
